moved linearity file
